rename responseActionCreator to requestActionCreator
update documentation to reflect naming change

update package.json to include peerDependencies properly
