+++ README.md
@@ -42,16 +42,16 @@ if you have multiple enhancers, simply using the redux <a href="https://redux.js
 ```typescript
 const store: ActionPromiseStore = createStore(myReducer, compose(ActionPromiseEnhancer, ...otherEnhancers));
 ```
-### Response Actions
+### Request Actions
 
-sometimes we want to dispatch an action, and we want to know when a following action has occurred where we dispatched our action, however, we can't easily do this with redux, this is where response actions come in, when you feed a response action to the dispatch function it will respond with a promise that you can await to get the payload of the subsequently dispatched action
+sometimes we want to dispatch an action, and we want to know when a following action has occurred where we dispatched our action, however, we can't easily do this with redux, this is where request actions come in, when you feed a request action to the dispatch function it will respond with a promise that you can await to get the payload of the subsequently dispatched action
 
 ```typescript
-import ActionPromiseEnhancer, { createResponseAction, ActionPromiseStore } from 'redux-action-promise-enhancer';
+import ActionPromiseEnhancer, { createRequestAction, ActionPromiseStore } from 'redux-action-promise-enhancer';
 
 const MyActionType1 = 'my-action';
 const MyResponseAction1 = {type: 'my-action-completed'};
-const responseActionCreator = createResponseAction((payload: number) => ({
+const requestActionCreator = createRequestAction((payload: number) => ({
         type: MyActionType1,
         payload
     }), MyResponseAction1);
@@ -59,7 +59,7 @@ const responseActionCreator = createResponseAction((payload: number) => ({
 const store: ActionPromiseStore = createStore(myReducer, ActionPromiseEnhancer);
 
 const dispatch = async () => {
-    const response = await store.dispatch(responseActionCreator(1));
+    const response = await store.dispatch(requestActionCreator(1));
     
     console.log('awaited response', response);
 };
@@ -82,20 +82,20 @@ request action {
 awaited response { type: 'my-action-completed' }
 ```
 
-handling errors in response actions
+handling errors in request actions
 
 ```typescript
-const MyActionType1 = 'my-action';
-const MyErrorAction1 = {type: 'my-action-error'};
-const responseAction = createResponseAction(
-    {type: MyActionType1},
+const MyRequestActionType1 = 'my-action';
+const MyResponseErrorAction1 = {type: 'my-action-error'};
+const requestAction = createRequestAction(
+    {type: MyRequestActionType1},
     undefined,
-    MyErrorAction1
+    MyResponseErrorAction1
 );
 
 const dispatch = async () => {
     try {
-        const response = await store.dispatch(responseAction);
+        const response = await store.dispatch(requestAction);
     
         console.log('awaited response', response);
     } catch (e) {
@@ -103,9 +103,9 @@ const dispatch = async () => {
     }
 };
 
-store.subscribeToActions(MyActionType1).addListener((action) => {
+store.subscribeToActions(MyRequestActionType1).addListener((action) => {
     console.log('request action', action)
-    store.dispatch(MyErrorAction1);
+    store.dispatch(MyResponseErrorAction1);
 });
 
 dispatch();
@@ -121,16 +121,16 @@ request action {
 error while executing { type: 'my-action-error' }
 ```
 
-timeout a response action
+timeout a request action
 
 ```typescript
-const responseAction = createResponseAction(
-    {type: MyActionType1}, 
+const requestAction = createRequestAction(
+    {type: MyRequestActionType1}, 
     MyResponseAction1,
     undefined,
     100
 );
-store.dispatch(responseAction)
+store.dispatch(requestAction)
     .catch((error: TimeoutError) => console.log(error.name, error.message));
 ```
 Logs `TimeoutError Timed out promise` after 100ms, the promise is reject with an `Error`
