+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "redux-action-promise-enhancer",
-  "version": "1.1.3",
+  "version": "1.2.0",
   "description": "Create a promise from a redux store and a list of expected actions that will resolve in the future",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -32,11 +32,12 @@
     "url": "https://github.com/vankovilija/redux-action-promise/issues"
   },
   "homepage": "https://github.com/vankovilija/redux-action-promise#readme",
-  "dependencies": {
-    "redux": "^4.0.5",
-    "typescript": "^4.1.3"
+  "peerDependencies": {
+    "redux": ">= 4.0.1 < 5"
   },
   "devDependencies": {
+    "redux": "^4.0.5",
+    "typescript": "^4.1.3",
     "@reduxjs/toolkit": "^1.5.0",
     "@types/jest": "^26.0.20",
     "jest": "^26.6.3",
