+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './action-promise-store.interface'
 export * from './reject-action-error'
 export * from './timeout-error'
-export * from './create-response-action'
+export * from './create-request-action'
 import { ActionPromiseEnhancer, ValidationMode } from './enhancer'
 
 export {
